Do not use absolute path to shared libraries that don't have a SONAME set
When there is no SONAME, the linker would embed the absolute path as DT_NEEDED in the
output, which will never work when using a sysroot, for example when cross-compiling.

This affected pkg-config dependencies, and find_library when the "dirs" argument is given.

Now if the library has no SONAME :
 - -L/-l arguments returned by pkg-config
 - find_library generates "-L/dir -lname", instead of "/dir/libname.so"

Fixes: #7766